requirements.txt
asgiref==3.3.1
boto3==1.16.53
botocore==1.19.53
Django==3.1.4
django-filter==2.4.0
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
